add functionality for redirecting unauthenticated users to the welcome page and a test case for it
