+++ core/settings.py
@@ -112,7 +112,8 @@ AUTH_PASSWORD_VALIDATORS = [
 
 AUTH_USER_MODEL = "forge.Worker"
 
-LOGIN_REDIRECT_URL = "/"
+LOGIN_REDIRECT_URL = "forge:index"
+LOGIN_URL = "forge:welcome"
 
 # Internationalization
 # https://docs.djangoproject.com/en/4.1/topics/i18n/
