+++ forge/urls.py
@@ -1,9 +1,10 @@
 from django.urls import path
 
-from core.views import index
+from forge.views import index, welcome
 
 urlpatterns = [
     path("", index, name="index"),
+    path("unregistered/", welcome, name="welcome")
 ]
 
 app_name = "forge"
