+++ forge/views.py
@@ -1,3 +1,30 @@
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponseRedirect
 from django.shortcuts import render
+from django.urls import reverse_lazy
+
+from forge.models import Worker
+
 
 # Create your views here.
+def welcome(request):
+    if request.user.is_authenticated:
+        return HttpResponseRedirect(reverse_lazy("forge:index"))
+    return render(request, "forge/unregistered/welcome.html")
+
+
+@login_required
+def index(request):
+    """View function for the home page of the site."""
+
+    num_workers = Worker.objects.count()
+
+    num_visits = request.session.get("num_visits", 0)
+    request.session["num_visits"] = num_visits + 1
+
+    context = {
+        "num_workers": num_workers,
+        "num_visits": num_visits,
+    }
+
+    return render(request, "forge/index.html", context=context)
